services/image-rec-manual-api/requirements.txt
aiohttp
aiohttp_cors
aioredis
protobuf>=3.6.1
python-socketio>=3.1.1
